refactor(tests): split oversized Python test files to comply with 500-line policy
Split 6 test files that exceeded 500 lines into focused modules:

- test_memory_tools.py (685 lines) → _crud, _permissions, _advanced
- test_relationship_memory.py (666 lines) → _core, _interactions
- test_persona_e2e.py (562 lines) → _grpc_events, _scheduling_memory
- test_working_memory.py (561 lines) → _sections, _compression
- test_relationship_memory_user.py (557 lines) → _core, _prompts
- test_builtin_tools.py (551 lines) → _filesystem_shell, _http

All shared imports, fixtures, and helpers are copied into each split
file so each is a valid standalone test module. No logic changes.

Removes the 6 grandfathered entries from GRANDFATHERED_FILES in
scripts/checks/file_size.py, and adds .claude/** to _EXTRA_EXCLUDES
to prevent worktree files from triggering false positives in the
file size hook.
