+++ FILEMAP.md
@@ -1,7 +1,7 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-25  
-> **445** tracked files
+> **452** tracked files
 
 ## Summary
 
@@ -9,8 +9,8 @@
 |---------------------|-------|
 | `internal/` | 126 |
 | `docs/` | 81 |
+| `tests/` | 78 |
 | `agents/` | 77 |
-| `tests/` | 71 |
 | `(root)` | 20 |
 | `scripts/` | 17 |
 | `.github/` | 15 |
@@ -26,7 +26,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 159 |
+| `.py` | 166 |
 | `.go` | 123 |
 | `.md` | 97 |
 | `.yaml` | 25 |
@@ -502,7 +502,8 @@
 │   │   ├── test_memory_budget_e2e.py
 │   │   ├── test_observability_e2e.py
 │   │   ├── test_observability_schema_parity.py
-│   │   ├── test_persona_e2e.py
+│   │   ├── test_persona_e2e_grpc_events.py
+│   │   ├── test_persona_e2e_scheduling_memory.py
 │   │   ├── test_span_links.py
 │   │   ├── test_trace_propagation.py
 │   │   └── test_workflow.py
@@ -516,7 +517,8 @@
 │           ├── test_agents.py
 │           ├── test_base_agent.py
 │           ├── test_base_handle.py
-│           ├── test_builtin_tools.py
+│           ├── test_builtin_tools_filesystem_shell.py
+│           ├── test_builtin_tools_http.py
 │           ├── test_defaults.py
 │           ├── test_dispatch_execute_actions.py
 │           ├── test_dispatch_lifecycle.py
@@ -535,7 +537,9 @@
 │           ├── test_memory_injection_advanced.py
 │           ├── test_memory_instructions.py
 │           ├── test_memory_notes.py
-│           ├── test_memory_tools.py
+│           ├── test_memory_tools_advanced.py
+│           ├── test_memory_tools_crud.py
+│           ├── test_memory_tools_permissions.py
 │           ├── test_participant.py
 │           ├── test_permissions.py
 │           ├── test_persona_agent_errors.py
@@ -545,8 +549,10 @@
 │           ├── test_persona_state.py
 │           ├── test_persona_timeouts.py
 │           ├── test_registration.py
-│           ├── test_relationship_memory.py
-│           ├── test_relationship_memory_user.py
+│           ├── test_relationship_memory_core.py
+│           ├── test_relationship_memory_interactions.py
+│           ├── test_relationship_memory_user_core.py
+│           ├── test_relationship_memory_user_prompts.py
 │           ├── test_sandbox.py
 │           ├── test_send_chat_message.py
 │           ├── test_server_channel.py
@@ -560,7 +566,8 @@
 │           ├── test_validate_agent_schema.py
 │           ├── test_validate_persona.py
 │           ├── test_validate_workflow.py
-│           └── test_working_memory.py
+│           ├── test_working_memory_compression.py
+│           └── test_working_memory_sections.py
 └── workflows/
     ├── budget-test.yaml
     └── feature-builder.yaml
